Remaining game C files: mechanical warning cleanup to zero

+++ src/actors/item_box/render.inc.c
@@ -28,7 +28,7 @@ void render_actor_item_box(Camera* camera, struct ItemBox* item_box) {
     f32 someMultiplier;
 
     size_t actorIdx = CM_FindActorIndex((struct Actor*)item_box);
-    if (-1 == actorIdx) {
+    if (-1 == (s32) actorIdx) {
         printf("[render_actor_item_box] Could not find item box for FI, skipping!\n");
         return;
     }
